migrate

+++ database/migrations/2023_05_04_155848_create_comments_table.php
@@ -15,8 +15,8 @@ class CreateCommentsTable extends Migration
     {
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->index();
-            $table->integer('document_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
+            $table->unsignedBigInteger('document_id')->index();
             $table->text('content');
             $table->timestamps();
         });
